Update Balance Storage Handling
Some system accounts like the FEE_RESERVE and the nToken can only hold positive cash balances. Getting and setting these balances is simplified in this commit.

For all other accounts, the getters and setters for balance storage are updated. _setBalanceStorage will convert the “signed prime supply value” stored on the stack and in memory to the proper prime cash storage value via convertToStorageNonSettlementNonVault.

Methods that use the balance handler getters are updated. Not all updates to balance setters are in this commit.