Adds Events to nToken Mint and Redeem
Refactors nToken mint and redeem code to emit proper Transfer events. nTokenRedeem has quite a few lines of changes, but these are primarily to deal with stack size issues. There are no intended business logic changes.

The erroneous “hasResidual” boolean is removed in favor of checking the newifCashAssets array length. This boolean was not set properly in prior code and required a hot fix.